llm-gateway/requirements.txt
fastapi>=0.104.1
httpx>=0.25.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
typing-extensions>=4.8.0
python-dotenv>=1.2.1